Moved the header title to context
